Move ImageWidgets code to generic file, WIP
